quest completed -> lvl 10 to lvl 479 -> + 1000000 coins

+++ src/main/java/model/bo/AvaliacaoBO.java
@@ -54,7 +54,7 @@ public class AvaliacaoBO implements GenericBO<Avaliacao> {
     }
 
     @Override
-    public Avaliacao getById(int id) throws Exception {
+    public Avaliacao getById(long id) throws Exception {
         genericDAO = new GenericDao<>();
         return genericDAO.getById(Avaliacao.class, id);
     }
