+++ src/main/java/model/bo/GenericBO.java
@@ -9,7 +9,7 @@ public interface GenericBO<T> {
     boolean deletar(T o) throws Exception;
     boolean alterar(T o) throws Exception;
     List<T> listarTodos() throws Exception;
-    T getById(int id) throws Exception;
+    T getById(long id) throws Exception;
     boolean valida(T o) throws Exception;
     boolean validaId(long id) throws Exception;
